+++ src/lua/LuaHandle.cxx
@@ -62,7 +62,7 @@ bool LuaHandle::devMode = false;
 //============================================================================//
 //============================================================================//
 
-LuaHandle::LuaHandle(const string& _name, int _scriptID,
+LuaHandle::LuaHandle(const string& _name, int16_t _scriptID,
 		                 int gameState, int drawWorld, int drawScreen,
                      bool _fullRead, bool _gameCtrl, bool _inputCtrl)
 : EventClient(_name, _scriptID,
